omega navbar functionality + pages structuring overhaul
