+++ package.json
@@ -14,6 +14,7 @@
     "@react-native-community/masked-view": "^0.1.11",
     "@react-navigation/bottom-tabs": "^6.3.1",
     "@react-navigation/native": "^6.0.10",
+    "@react-navigation/native-stack": "^6.6.2",
     "@react-navigation/stack": "^6.2.1",
     "expo": "~44.0.0",
     "expo-dev-client": "~0.8.5",
@@ -31,7 +32,8 @@
     "react-native-screens": "~3.10.1",
     "react-native-simple-time-picker": "^1.3.11",
     "react-native-svg": "^12.3.0",
-    "react-native-web": "0.17.1"
+    "react-native-web": "0.17.1",
+    "react-navigation": "^4.4.4"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
